Fetch from origin

+++ ZXingObjC/pdf417/ZXPDF417Common.h
@@ -33,10 +33,9 @@ extern const int ZX_PDF417_MODULES_IN_STOP_PATTERN;
 + (ZXIntArray *)toIntArray:(NSArray *)list;
 
 /**
- * Translate the symbol into a codeword.
- *
+ * @param symbol encoded symbol to translate to a codeword
  * @return the codeword corresponding to the symbol.
  */
-+ (int)codeword:(long)symbol;
++ (int)codeword:(int)symbol;
 
 @end
